fix(root): move Audio/Screen to common/entities and remove 10 dead handler-manager imports to break root<->handlers cycle
