+++ apps/aurora-core/src/modules/common/entities/audio.ts
@@ -1,5 +1,5 @@
 import { Entity } from 'typeorm';
-import SubscribeEntity from '../../common/entities/subscribe-entity';
+import SubscribeEntity from './subscribe-entity';
 
 @Entity()
 export default class Audio extends SubscribeEntity {
