+++ apps/aurora-core/src/modules/common/entities/index.ts
@@ -1,2 +1,4 @@
 export { default as BaseEntity } from './base-entity';
 export { default as SubscribeEntity } from './subscribe-entity';
+export { default as Audio } from './audio';
+export { default as Screen } from './screen';
