+++ apps/aurora-core/src/modules/root/entities/index.ts
@@ -1,9 +1,9 @@
 // eslint-disable-next-line import/no-cycle -- TODO fix cyclic dependency
-import Audio from './audio';
-import Screen from './screen';
+import Audio from '../../common/entities/audio';
+import Screen from '../../common/entities/screen';
 
-export { default as Audio } from './audio';
+export { default as Audio } from '../../common/entities/audio';
 export { BaseEntity, SubscribeEntity } from '../../common/entities';
-export { default as Screen } from './screen';
+export { default as Screen } from '../../common/entities/screen';
 
 export const Entities = [Audio, Screen];
