+++ apps/aurora-core/src/modules/root/handler-manager.ts
@@ -2,25 +2,14 @@ import { Server } from 'socket.io';
 import BaseScreenHandler from '../handlers/base-screen-handler';
 import type BaseHandler from '../handlers/base-handler';
 import SubscribeEntity from '../common/entities/subscribe-entity';
-import SimpleAudioHandler from '../handlers/audio/simple-audio-handler';
 import dataSource from '../../database';
 import { Audio, Screen } from './entities';
 import { LightsGroup } from '../lights/entities';
-import { RandomEffectsHandler } from '../handlers/lights';
-import SetEffectsHandler from '../handlers/lights/set-effects-handler';
-import DevelopEffectsHandler from '../handlers/lights/develop-effects-handler';
 import { BeatEvent, TrackChangeEvent } from '../events/music-emitter-events';
-import { CurrentlyPlayingTrackHandler, CenturionScreenHandler } from '../handlers/screen';
-// eslint-disable-next-line import/no-cycle -- TODO fix cyclic dependency
-import { ScenesHandler } from '../handlers/lights/scenes-handler';
-import EffectSequenceHandler from '../handlers/lights/effect-sequence-handler';
 import { MusicEmitter } from '../events';
-import StageEffectsHandler from '../handlers/screen/stage-effects-handler';
 import { SocketioNamespaces } from '../../socketio-namespaces';
 import logger from '../../logger';
 import { BackofficeSyncEmitter } from '../events/backoffice-sync-emitter';
-import TimeTrailRaceScreenHandler from '../handlers/screen/time-trail-race-screen-handler';
-import TimeTrailRaceLightsHandler from '../handlers/lights/time-trail-race-lights-handler';
 import HandlerFactory from './handler-factory';
 import { ShowOrdersEvent } from '../events/order-emitter-events';
 import EmitterStore from '../events/emitter-store';
